requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
asgiref==3.9.1
asyncio==3.4.3
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.7.9
charset-normalizer==3.4.2
discord.py==2.5.2
dj-database-url==3.0.1
dj-email-url==1.0.6
Django==5.2.4
django-cache-url==3.4.5
djangorestframework==3.16.0
environs==14.2.0
frozenlist==1.7.0
gunicorn==23.0.0
idna==3.10
marshmallow==4.0.0
multidict==6.6.3
packaging==25.0
propcache==0.3.2
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
yarl==1.20.1
